requirements.txt
Django>=1.7.1
django-betty-cropper>=0.1.4
django-bulbs==0.4.6
